gnn-llm/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
pinecone-client>=2.2.4
sentence-transformers>=2.2.2
networkx>=3.1
numpy>=1.24.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
psutil>=5.9.0
schedule>=1.2.0